[scripts] arm: execute profiles update outside qemu-user chroot

+++ scripts/inner_source_chroot_update.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+TARGET="${1}"
+
 /usr/sbin/env-update
 . /etc/profile
 
@@ -118,14 +120,17 @@ for conf in 00-sabayon.package.use 00-sabayon.package.mask \
 	fi
 done
 
-# Update /usr/portage/profiles
-# This is actually not strictly needed but several
-# gentoo tools expect to find valid /etc/make.profile symlink
-# This part is best effort, if it will be able to complete
-# correctly, fine.
-# For a list of mirrors, see: http://www.gentoo.org/main/en/mirrors-rsync.xml
-RSYNC_URI="rsync://rsync.at.gentoo.org/gentoo-portage/profiles"
-PROFILES_DIR="/usr/portage/profiles"
-safe_run rsync -av -H -A -X --delete-during "${RSYNC_URI}/" "${PROFILES_DIR}/"
+# qemu-user arm has problems with rsync, deadlocks?
+if [ "${TARGET}" != "arm" ]; then
+	# Update /usr/portage/profiles
+	# This is actually not strictly needed but several
+	# gentoo tools expect to find valid /etc/make.profile symlink
+	# This part is best effort, if it will be able to complete
+	# correctly, fine.
+	# For a list of mirrors, see: http://www.gentoo.org/main/en/mirrors-rsync.xml
+	RSYNC_URI="rsync://rsync.at.gentoo.org/gentoo-portage/profiles"
+	PROFILES_DIR="/usr/portage/profiles"
+	safe_run rsync -av -H -A -X --delete-during "${RSYNC_URI}/" "${PROFILES_DIR}/"
+fi
 
 equo query list installed -qv > /etc/sabayon-pkglist
